### 2017-12-06 #483 @whedon
[PRE REVIEW]: kima: Exoplanet detection in radial velocities
**Submitting author:** @j-faria (<a href="http://orcid.org/0000-0002-6728-244X">João Faria</a>)
**Repository:** <a href="https://github.com/j-faria/kima" target ="_blank">https://github.com/j-faria/kima</a>
**Version:** v1.0.0
