### 2017-12-06 #483 @arfon
👋 @barentsen - would you be interested in reviewing this submission for JOSS?
