### 2017-12-06 #483 @barentsen
@arfon It would be a privilege to review this package, however I will likely not get to it until next Monday due to travel and commitments.  Would that be acceptable?
