+++ Makefile
@@ -8,7 +8,7 @@ EIGEN_PATH = eigen
 includes = -I$(DNEST4_PATH) -I$(EIGEN_PATH) 
 # -I$(CELERITE_PATH)
 
-CXX = g++
+CXX = g++-4.9
 CXXFLAGS = -pthread -std=c++11 -O3 -DNDEBUG -w -DEIGEN_MPL2_ONLY
 LIBS = -ldnest4 -L/usr/local/lib
 
@@ -69,9 +69,8 @@ clean:
 	rm -f main $(OBJS)
 
 cleanout:
-	rm -f sample.txt sample_info.txt levels.txt weights.txt posterior_sample.txt sampler_state.txt
+	rm -f sample.txt sample_info.txt levels.txt \
+              weights.txt posterior_sample.txt sampler_state.txt \
+              posterior_sample_lnlikelihoods.txt
 
-cleanall: clean
-	rm -f sample.txt sample_info.txt \
-		levels.txt weights.txt posterior_sample.txt sampler_state.txt \
-		posterior_sample_lnlikelihoods.txt
+cleanall: clean cleanout
