+++ Makefile
@@ -8,7 +8,7 @@ EIGEN_PATH = eigen
 includes = -I$(DNEST4_PATH) -I$(EIGEN_PATH) 
 # -I$(CELERITE_PATH)
 
-CXX = g++-4.9
+CXX = g++
 CXXFLAGS = -pthread -std=c++11 -O3 -DNDEBUG -w -DEIGEN_MPL2_ONLY
 LIBS = -ldnest4 -L/usr/local/lib
 
