### 2017-12-08 0b92e72
specify std::abs
