### 2017-12-10 2fb0c2f
complete output description; should fix #17
