### 2017-12-10 dc1cd16
maybe a better way to update the prior; related to #15
