+++ src/RVConditionalPrior.cpp
@@ -12,45 +12,12 @@ extern ContinuousDistribution *log_muP_prior;
 extern ContinuousDistribution *wP_prior;
 extern ContinuousDistribution *log_muK_prior;
 
+extern ContinuousDistribution *Pprior;
+extern ContinuousDistribution *Kprior;
 
-/*
-if(hyperpriors)
-{
-    // Cauchy prior centered on log(365 days), scale=1
-    // for log(muP), with muP in days
-    // -- truncated to (~-15.1, ~26.9) --
-    TruncatedCauchy log_muP_prior(log(365), 1., log(365)-21, log(365)+21);
-
-    // Uniform prior between 0.1 and 3
-    // for wP
-    Uniform wP_prior(0.1, 3.);
-    
-    // Cauchy prior centered on log(1), scale=1
-    // for log(muK), with muK in m/s
-    // -- truncated to (-21, 21) --
-    // NOTE: we actually sample on muK itself, just the prior is for log(muK). 
-    // Confusing, I know...
-    TruncatedCauchy log_muK_prior(0., 1., 0.-21, 0.+21);
-
-    Laplace Pprior;
-    Exponential Kprior;
-
-}
-else
-{
-*/
-    //extern Jeffreys Pprior;
-    extern ContinuousDistribution *Pprior;
-    extern ContinuousDistribution *Kprior;
-    // Jeffreys Pprior(1.0, 1E7); // days
-    //ModifiedJeffreys Kprior(1.0, 1E4); // m/s
-    //Uniform Kprior(0., 20.); // m/s
-
-    extern ContinuousDistribution *eprior;
-    extern ContinuousDistribution *phiprior;
-    extern ContinuousDistribution *wprior;
-    //TruncatedRayleigh eprior(0.2, 0.0, 1.0);
-    //TruncatedNormal eprior(0, 0.3, 0., 1.);
+extern ContinuousDistribution *eprior;
+extern ContinuousDistribution *phiprior;
+extern ContinuousDistribution *wprior;
 
 
 RVConditionalPrior::RVConditionalPrior()
@@ -60,7 +27,6 @@ RVConditionalPrior::RVConditionalPrior()
 
 void RVConditionalPrior::from_prior(RNG& rng)
 {
-    //cout << "called RVConditionalPrior::from_prior !!!" << endl;
     if(hyperpriors)
     {
         center = log_muP_prior->rvs(rng);
@@ -78,9 +44,7 @@ double RVConditionalPrior::perturb_hyperparameters(RNG& rng)
         int which = rng.rand_int(3);
 
         if(which == 0)
-        {
             log_muP_prior->perturb(center, rng);
-        }        
         else if(which == 1)
             wP_prior->perturb(width, rng);
         else
@@ -102,16 +66,6 @@ double RVConditionalPrior::perturb_hyperparameters(RNG& rng)
 
 double RVConditionalPrior::log_pdf(const std::vector<double>& vec) const
 {
-    // cout << "type of Pprior:" << typeid(Pprior).name() << endl;
-    //cout << "called RVConditionalPrior::log_pdf !!!" << endl;
-
-    // Determine whether to use the global priors for P/K
-    // or whether to use ones based on the hyperparameters
-    Laplace l(center, width);
-    Exponential e(muK);
-    ContinuousDistribution* _Pprior = Pprior;
-    ContinuousDistribution* _Kprior = Kprior;
-
     if(hyperpriors)
     {
         if(vec[2] < 0. || vec[2] > 2.*M_PI ||
@@ -119,13 +73,8 @@ double RVConditionalPrior::log_pdf(const std::vector<double>& vec) const
[…]
