### 2017-12-15 154304e
better prior for eta4
