+++ src/default_priors.h
@@ -24,7 +24,7 @@ Uniform *wprior = new Uniform(0.0, 2*M_PI);
 Uniform *log_eta1_prior = new Uniform(-5, 5);
 Uniform *log_eta2_prior = new Uniform(0, 5);
 Uniform *eta3_prior = new Uniform(10., 40.);
-Uniform *log_eta4_prior = new Uniform(-5, 0);
+Uniform *log_eta4_prior = new Uniform(-1, 1);
 
 
 
