### 2017-12-18 12dd8d0
change acknowledgement references; clean up
