### 2017-12-18 89aed0f
optional arguments to set model options and correct order; deal with no planets
