### 2017-12-20 #487 @barentsen
I expect to carry out this review on Dec 23-24.  Apologies for the delay!
