+++ scripts/display.py
@@ -52,7 +52,7 @@ def percentile68_ranges_latex(a):
 
 
 def get_planet_mass(P, K, e, star_mass=1.0, full_output=False, verbose=False):
-    if verbose: print 'Using star mass = %s solar mass' % star_mass
+    if verbose: print('Using star mass = %s solar mass' % star_mass)
     # print 3.5e-2 * K * (P/365.)**(1./3) * star_mass**(2./3)
     # print 9.077e-3 * star_mass**(2./3) * (P/(2*np.pi))**(1./3) * K * np.sqrt(1-e**2)
 
@@ -88,10 +88,10 @@ class DisplayResults(object):
         top_level = os.path.dirname(os.path.dirname(path_to_this_file))
 
         if debug:
-            print 
-            print 'running on:', pwd
-            print 'top_level:', top_level
-            print 
+            print()
+            print('running on:', pwd)
+            print('top_level:', top_level)
+            print()
 
         def get_skip(line):
             load_args = re.findall(r'\((.*?)\)', line, re.DOTALL)[1]
@@ -126,10 +126,10 @@ class DisplayResults(object):
                 data_file = re.findall('"(.*?)"', line, re.DOTALL)[0]
                 data_file = pathjoin(top_level, data_file)
 
-        print 'Loading data file %s' % data_file
+        print('Loading data file %s' % data_file)
         self.data_file = data_file
         if debug:
-            print '--- skipping first %d rows of data file' % self.data_skip
+            print('--- skipping first %d rows of data file' % self.data_skip)
 
         self.data = np.loadtxt(self.data_file, 
                                skiprows=self.data_skip, usecols=(0,1,2))
@@ -152,7 +152,7 @@ class DisplayResults(object):
             self.max_likelihood_index = np.argmax(self.posterior_sample_lnlikes)
         
         except IOError:
-            print 'Sample likelihoods not available!!! This is bad!'
+            print('Sample likelihoods not available!!! This is bad!')
 
 
         start_parameters = 0
@@ -167,7 +167,7 @@ class DisplayResults(object):
                 self.trend = 'bool trend = true' in f.read()
         
         if debug: 
-            print 'trend:', self.trend
+            print('trend:', self.trend)
 
 
         if self.trend:
@@ -194,7 +194,7 @@ class DisplayResults(object):
             self.fiber_offset = fiber_offset
 
         if debug: 
-            print 'obs_after_fibers:', self.fiber_offset
+            print('obs_after_fibers:', self.fiber_offset)
 
         if self.fiber_offset:
             n_offsets = 1
@@ -213,7 +213,7 @@ class DisplayResults(object):
                 self.GPmodel = 'bool GP = true' in f.read()
         
         if debug:
-            print 'GP model:', self.GPmodel
+            print('GP model:', self.GPmodel)
 
         if self.GPmodel:
             n_hyperparameters = 4
@@ -330,7 +330,7 @@ class DisplayResults(object):
     def get_medians(self):
         """ return the median values of all the parameters """
         if self.posterior_sample.shape[0] % 2 == 0:
-            print 'Median is not a solution because number of samples is even!!'
+            print('Median is not a solution because number of samples is even!!')
 
         self.medians = np.median(self.posterior_sample, axis=0)
[…]
