### 2017-12-24 j-faria/kima#18 @barentsen
The `many_planets` example does not compile due to missing header files
The `many_planets` example does not compile because three header files appear to be missing from the repository: https://github.com/j-faria/kima/blob/12dd8d0a387416d719c31af5735d833723e57b25/examples/many_planets/kima_setup.cpp#L12-L14

(This issue is part of https://github.com/openjournals/joss-reviews/issues/487)
