### 2017-12-24 j-faria/kima#19 @barentsen
All three example fail to compile due to recent API change (objects -> planets)
All three examples (`BL2009`, `CoRoT7`, `many_planets`) yield an error message during compilation (appened below).  It looks like the error is caused by the recent renaming of `RVmodel():objects` into `RVmodel():planets` (cf. commit ffbc79264644c853abd79e0fdcf0086b5ea40ba2).

This issue will be easy to fix, but it demonstrates that this software will benefit from having the examples tested automatically upon each change (e.g. using Travis).  I will be opening a separate issue about this.

```Makefile
~/dev/kima/examples/CoRoT7 $ make
g++ -c -I../../src -I../../DNest4/code -I../../eigen -I../../celerite/cpp/include -o kima_setup.o kima_setup.cpp -pthread -std=c++11 -O3 -w
kima_setup.cpp: In constructor ‘RVmodel::RVmodel()’:
kima_setup.cpp:22:6: error: class ‘RVmodel’ does not have any field named ‘objects’
     :objects(5, 5, false, RVConditionalPrior())
      ^
kima_setup.cpp:24:58: error: no matching function for call to ‘DNest4::RJObject<RVConditionalPrior>::RJObject()’
     ,C(Data::get_instance().N(), Data::get_instance().N())
                                                          ^
In file included from ../../DNest4/code/RJObject/RJObject.h:129:0,
                 from ../../DNest4/code/DNest4.h:18,
                 from kima_setup.cpp:2:
../../DNest4/code/RJObject/RJObjectImpl.h:2:1: note: candidate: DNest4::RJObject<ConditionalPrior>::RJObject(int, int, bool, const ConditionalPrior&, DNest4::PriorType) [with ConditionalPrior = RVConditionalPrior]
 RJObject<ConditionalPrior>::RJObject(int num_dimensions, int max_num_components, bool fixed,
 ^
../../DNest4/code/RJObject/RJObjectImpl.h:2:1: note:   candidate expects 5 arguments, 0 provided
In file included from ../../DNest4/code/DNest4.h:18:0,
                 from kima_setup.cpp:2:
../../DNest4/code/RJObject/RJObject.h:27:7: note: candidate: DNest4::RJObject<RVConditionalPrior>::RJObject(const DNest4::RJObject<RVConditionalPrior>&)
 class RJObject
       ^
../../DNest4/code/RJObject/RJObject.h:27:7: note:   candidate expects 1 argument, 0 provided
../../DNest4/code/RJObject/RJObject.h:27:7: note: candidate: DNest4::RJObject<RVConditionalPrior>::RJObject(DNest4::RJObject<RVConditionalPrior>&&)
../../DNest4/code/RJObject/RJObject.h:27:7: note:   candidate expects 1 argument, 0 provided
Makefile:23: recipe for target 'kima_setup.o' failed
make: *** [kima_setup.o] Error 1
```

(This issue is part of openjournals/joss-reviews#487)
