### 2017-12-24 j-faria/kima#19 @j-faria
Thanks, @barentsen. Indeed this particular error is an easy fix, as you noted, but I also agree the long-term solution is to build the examples in Travis. I will take care of that ASAP.
