### 2017-12-24 #487 @barentsen
I am opening a few issues on the repository now because I happen to have a bit of spare time, but, needless to say: I do *not* expect a quick reply! @j-faria please enjoy your holidays and don't let my activity trigger you to work over the xmas-nye period, we all need time off to preserve mental health! =)
