### 2017-12-25 j-faria/kima#20 @barentsen
Turn the `scripts` directory into a Python package
A key component of `kima` is the analysis code located under `scripts`, which contains a significant number of Python functions and classes used to interpret and analyze results. Because this code is re-used in all the examples, and because you encourage users to re-use it as well, I strongly encourage you to turn this directory into a Python package.

Creating a Python package is surprisingly easy!  It essentially takes three steps:
1. Rename `scripts` into `kima` (or however you want to name the Python package).
2. Add an empty `__init__.py` file inside `kima` (this is what turns a directory into a Python package).
3. Add a `setup.py` file to the repository to contain the metadata and make it installable (i.e. the docs would then instruct the user to type `python setup.py install`).

Doing this would make it much easier to re-use the Python code, e.g. users will be able to use `import kima.dnest4` or `import kima.display` from any location.

Right now, it looks like users are requested to copy code from `scripts` to a custom work directory.  Because the `showresults.py` script requires modules from `scripts`, I see that the code has been modified in the examples directory to include a `sys.path.append()` statement.  This hard-coded method of ensuring that the required modules can be imported is error-prone, and the docs do not explain it in sufficient detail.  Instead I recommend creating a package, it is the standard way to enable re-use of Python functions and classes.

Two extra tips:
* An extra advantage of a package is that you can configure the `setup.py` file to add command-line tools to the user's path upon installation.  For example, if you add the option `entry_points={'console_scripts': ['kima-showresults = kima.showresults:showresults_main']}`, then installing kima will add a command-line tool called `kima-showresults` which executes the function `kima.showresults.showresults_main` when a user runs it on the command line.
* Once you have created the package, you can install it using `python setup.py develop`.  This will install the package using symlinks to the development directory so that you don't have to type `python setup.py install` after each change.  It makes developing a package more fun.

If this sounds like a good idea but you get stuck while creating the setup.py file, then do shout for help!

(This issue is part of https://github.com/openjournals/joss-reviews/issues/487)
