### 2017-12-25 j-faria/kima#21 @barentsen
Add automated tests to verify the core functionality
I strongly encourage the authors to add automated tests to enable users to verify whether their installation of kima works correctly, and to help prevent future code changes from introducing bugs (e.g. #18 could have been avoided this way).

It is great to see that the repository already contains a `.travis.yml` file which verifies that the C++ code can be compiled.  This is a great and important first step, but the package could be improved by adding tests which also verify the correct functionality of the code.

In Python this is really easy thanks to the `py.test` package.  For example, you can create a file called `tests/test_kima.py` which contains one or more functions with the prefix `test_` that execute `assert` statements. Then, when you run `py.test` on the command line, it will automatically discover those tests and tell you whether any of the assert statements evaluate to False.

An alternative to automated tests could be to include additional input/output in the docs to allow users to manually verify the correctness.  Some output is already shown in the docs (at https://github.com/j-faria/kima/wiki/5.-Analysis-of-results), but it is not clear what the exact input was.

(This issue is part of https://github.com/openjournals/joss-reviews/issues/487.)
