### 2017-12-25 j-faria/kima#23 @barentsen
Consider using the `dnest4` Python bindings
Just a suggestion: have you considered using the Python bindings for `dnest4`, or otherwise creating a Python wrapper for your own C++ code?

This could simplify this package by bypassing the need for compilation instructions, and bypassing the need for documenting both the C++ and the Python code.  It would also make it easier to treat `dnest4` as a standard Python dependency (i.e. it could automatically be installed and compiled via `pip` or `setup.py` just like e.g. astropy).

(This suggestion is part of https://github.com/openjournals/joss-reviews/issues/487.)
