### 2017-12-25 j-faria/kima#24 @barentsen
Cite and acknowledge other RV modeling packages
In both the JOSS paper and the README/docs, please acknowledge and cite other software packages that enable similar RV analyses.  In particular, it is important to clarify exactly what the unique contribution of `kima` is compared to those other packages.

For example: is kima the first tool to use nested sampling?  Is it the first to use GPs?  Is it the first to allow the number of planets to vary? Is it faster? Is it more accurate? Are any of these features important? etc, etc...

I'm sure you can think of some key advantages offered by kima.  It is important to state clearly what they are.  I have opened this issues because [the guidelines from the Journal of Open Source Software](https://joss.theoj.org/about) include the following two requirements:

> [The submission] should be a significant contribution to the available open source software that either enables some new research challenges to be addressed or makes addressing research challenges significantly better (e.g., faster, easier, simpler)

> Submissions that implement solutions already solved in other software packages are accepted into JOSS provided that they meet the criteria listed above and cite prior similar work.

FYI: Here are the first five alternative packages which I find via Google:
https://github.com/California-Planet-Search/radvel
https://github.com/nespinoza/exonailer
https://github.com/jdeast/EXOFASTv2
https://github.com/oscaribv/pyaneti
https://github.com/mrtommyb/ktransit

(This is part of https://github.com/openjournals/joss-reviews/issues/487.)
