### 2017-12-25 #487 @barentsen
I installed and ran `kima` and identified four key issues which I would like to see addressed before continuing my review:
- [ ] The examples do not compile (https://github.com/j-faria/kima/issues/18 & https://github.com/j-faria/kima/issues/19).
- [ ] There are no tests (be it manual or automated) which can be used to verify the core functionality (https://github.com/j-faria/kima/issues/21).
- [ ] The package does not follow the best practices in terms of code re-use/duplication; I encourage the authors to turn the `scripts` directory into a Python package (https://github.com/j-faria/kima/issues/20).
- [ ] The paper/docs do not explain in sufficient detail how the package compares to similar software tools (https://github.com/j-faria/kima/issues/24).

I do think astronomy would benefit from having a high-quality open source software tool which contains the features and ideas provided by `kima` (e.g. nested sampling and GPs), so I strongly encourage the authors to address these issues.  I am happy to answer further questions and mentor the improvements; i.e. I would love to see this package accepted eventually!
