### 2017-12-28 #487 @j-faria
Hi @barentsen, thank you for your thorough and very helpful review of `kima`!
I agree these four points you identified are necessary and useful for the package so I'll make the changes and add the features in the next few days.

I sincerely appreciate your message about work during the holiday period. For a PhD student like me, such a balance is an important and often forgotten skill that needs mastering, so it's nice to see it highlighted.
