### 2017-12-28 07352be
rename objects->planets in the examples; should fix #19
