+++ examples/BL2009/kima_setup.cpp
@@ -22,7 +22,7 @@ const bool trend = false;
 // options for the model
 // 
 RVmodel::RVmodel()
-    :objects(5, 1, false, RVConditionalPrior())
+    :planets(5, 1, false, RVConditionalPrior())
     ,mu(Data::get_instance().N())
     ,C(Data::get_instance().N(), Data::get_instance().N())
 {
