### 2017-12-28 d6a1e3b
compile the examples during the build; related to #19
