+++ .travis.yml
@@ -11,7 +11,7 @@ before_install:
   - if [[ "$TRAVIS_OS_NAME" == "osx" ]]; then brew update          ; fi
 
 script:
-  - make
+  - make -j
 
 matrix:
   allow_failures:
