### 2017-12-28 bba6662
add missing header files with default priors; should fix #18
