### 2017-12-30 6fd5942
first attempt at the pykima package (#20)
