+++ pykima/display.py
@@ -1,0 +1,883 @@
+import sys
+import re
+import os
+pathjoin = os.path.join
+
+from .keplerian import keplerian
+
+import matplotlib.pyplot as plt
+import numpy as np
+from scipy.stats import gaussian_kde
+import corner
+
+try:
+    from fast_histogram import histogram1d, histogram2d
+    fast_histogram_available = True
+except ImportError:
+    fast_histogram_available = False
+
+try:
+    from astroML.plotting import hist_tools
+    hist_tools_available = True
+except ImportError:
+    hist_tools_available = False
+
+
+colors = ["#9b59b6", "#3498db", "#95a5a6", "#e74c3c", "#34495e", "#2ecc71"]
+mjup2mearth = 317.8284065946748
+
+
+def apply_argsort(arr1, arr2, axis=-1):
+    """
+    Apply arr1.argsort() on arr2, along `axis`.
+    """
+    # check matching shapes
+    assert arr1.shape == arr2.shape, "Shapes don't match!"
+
+    i = list(np.ogrid[[slice(x) for x in arr1.shape]])
+    i[axis] = arr1.argsort(axis)
+    return arr2[i]
+
+def percentile68_ranges(a):
+    lp, median, up = np.percentile(a, [16, 50, 84])
+    return (median, up-median, median-lp)
+
+def percentile68_ranges_latex(a):
+    lp, median, up = np.percentile(a, [16, 50, 84])
+    return r'$%.2f ^{+%.2f} _{-%.2f}$' % (median, up-median, median-lp)
+
+# def get_aliases(Preal):
+#     fs = np.array([0.0027381631, 1.0, 1.0027]) #, 0.018472000025212765])
+#     return np.array([abs(1 / (1./Preal + i*fs)) for i in range(-2, 2)]).T
+
+
+def get_planet_mass(P, K, e, star_mass=1.0, full_output=False, verbose=False):
+    if verbose: print('Using star mass = %s solar mass' % star_mass)
+    # print 3.5e-2 * K * (P/365.)**(1./3) * star_mass**(2./3)
+    # print 9.077e-3 * star_mass**(2./3) * (P/(2*np.pi))**(1./3) * K * np.sqrt(1-e**2)
+
+    if isinstance(P, float):
+        assert isinstance(star_mass, float)
+        m_mj = 4.919e-3 * star_mass**(2./3) * P**(1./3) * K * np.sqrt(1-e**2)
+        m_me = m_mj * mjup2mearth
+        return m_mj, m_me
+    else:
+      if isinstance(star_mass, tuple) or isinstance(star_mass, list):
+        star_mass = star_mass[0] + star_mass[1]*np.random.randn(P.size)
+      m_mj = 4.919e-3 * star_mass**(2./3) * P**(1./3) * K * np.sqrt(1-e**2)
+      m_me = m_mj * mjup2mearth
+      
+      if full_output:
+        # return map(percentile68_ranges, [m_mj, m_me]), m_mj
+        return m_mj.mean(), m_mj.std(), m_mj
+      else:
+        # return map(percentile68_ranges, [m_mj, m_me])
+        return (m_mj.mean(), m_mj.std(), m_me.mean(), m_me.std())
+
+
+class KimaResults(object):
+    def __init__(self, options, data_file=None, 
+                 fiber_offset=None, hyperpriors=None,
+                 posterior_samples_file='posterior_sample.txt'):
+
+        self.options = options
+        debug = 'debug' in self.options
+
+        pwd = os.getcwd()
+        path_to_this_file = os.path.abspath(__file__)
+        top_level = os.path.dirname(os.path.dirname(path_to_this_file))
+
+        if debug:
+            print()
+            print('running on:', pwd)
+            print('top_level:', top_level)
[…]
