### 2017-12-30 0079507
install pykima in travis
