### 2017-12-30 f9b33c9
miniconda in travis with some (not all) dependencies
