+++ .travis.yml
@@ -1,5 +1,8 @@
 language: cpp
 
+python:
+  - 2.7
+
 os:
   - linux
   - osx
@@ -9,9 +12,18 @@ compiler:
 
 before_install:
   - if [[ "$TRAVIS_OS_NAME" == "osx" ]]; then brew update          ; fi
+  # setup miniconda
+  - wget http://repo.continuum.io/miniconda/Miniconda-latest-Linux-x86_64.sh -O miniconda.sh
+  - chmod +x miniconda.sh
+  - ./miniconda.sh -b
+  - export PATH=/home/travis/miniconda/bin:$PATH
+  - conda update --yes conda
+  - echo $TRAVIS_PYTHON_VERSION
 
 script:
   - make -j
+  # now the pykima package
+  - conda install --yes python=$TRAVIS_PYTHON_VERSION numpy scipy matplotlib
   - python setup.py install
 
 matrix:
