### 2017-12-30 45ee277
add argument parser to showresults
