### 2017-12-30 j-faria/kima#20 @j-faria
This was an excellent suggestion, so I created the `pykima` package, together with the `kima-showresults` command-line tool (which now hopefully takes better care of its arguments 45ee277f2c4275d87c4f1af5e5d8affc127594b4)

Travis is now running `python setup.py install` on every build, but the Python dependencies are still not correctly checked. Also, I need to update the documentation.
Once that's done, I will consider closing this issue.
