### 2017-12-30 e401f77
improve makefile
