### 2018-01-03 cbaa520
cleanout in examples' makefile
