+++ examples/many_planets/Makefile
@@ -8,7 +8,7 @@ CELERITE_PATH = $(KIMA_DIR)/celerite/cpp/include
 includes = -I$(SRC_DIR) -I$(DNEST4_PATH) -I$(EIGEN_PATH) -I$(CELERITE_PATH)
 
 CXX = g++
-CXXFLAGS = -pthread -std=c++11 -O3 -w -g
+CXXFLAGS = -pthread -std=c++11 -O3 -w
 LIBS = -ldnest4 -L/usr/local/lib
 
 KIMA_SRCS =\
@@ -28,7 +28,9 @@ kima: $(KIMA_OBJS)
 clean:
 	rm -f kima_setup.o kima
 
-cleanall: clean
+cleanout:
 	rm -f sample.txt sample_info.txt \
 		levels.txt weights.txt posterior_sample.txt sampler_state.txt \
 		posterior_sample_lnlikelihoods.txt
+
+cleanall: clean cleanout
