### 2018-01-03 d66f46b
add tests of pykima.keplerian and of default prior example
