+++ tests/conftest.py
@@ -1,0 +1,13 @@
+import pytest
+def pytest_addoption(parser):
+    parser.addoption("--slow", action="store_true",
+                     default=False, help="run slow tests")
+
+def pytest_collection_modifyitems(config, items):
+    if config.getoption("--slow"):
+        # --runslow given in cli: do not skip slow tests
+        return
+    skip_slow = pytest.mark.skip(reason="need --slow option to run")
+    for item in items:
+        if "slow" in item.keywords:
+            item.add_marker(skip_slow)
