### 2018-01-03 19de674
add example that samples from the default priors
