### 2018-01-03 0852f2a
add default_priors test to makefile; run tests in travis
