+++ .travis.yml
@@ -8,15 +8,19 @@ compiler:
   - gcc
 
 before_install:
-  - if [[ "$TRAVIS_OS_NAME" == "osx" ]]; then brew update          ; fi
+  - if [[ "$TRAVIS_OS_NAME" == "osx" ]]; then brew update; fi
   # setup miniconda
-  - wget http://repo.continuum.io/miniconda/Miniconda-latest-Linux-x86_64.sh -O ~/miniconda.sh
+  - if [[ "$TRAVIS_OS_NAME" == "osx" ]]; then
+      wget http://repo.continuum.io/miniconda/Miniconda-latest-MacOSX-x86_64.sh -O ~/miniconda.sh;
+    else
+      wget http://repo.continuum.io/miniconda/Miniconda-latest-Linux-x86_64.sh -O ~/miniconda.sh;
+    fi
   - bash ~/miniconda.sh -b -p $HOME/miniconda
   - export PATH="$HOME/miniconda/bin:$PATH"
   - conda update --yes conda
   - python --version
   - pip --version
-  - conda install --yes numpy scipy matplotlib pandas 
+  - conda install --yes numpy scipy matplotlib pandas pytest
 
 script:
   - make -j
