### 2018-01-03 j-faria/kima#21 @j-faria
A few tests added with 19de674d9838fbf5bf619b19a70097cf185ef2e3 and d66f46b5686a31281dedb1aa0bf59f032af901b0.
Will add this to the automated Travis builds
