### 2018-01-03 cc452c4
more tests in travis
