+++ .travis.yml
@@ -25,8 +25,10 @@ before_install:
 script:
   - make -j
   # now the pykima package
-  # - python setup.py install
-  - pip install -e .
+  - which python
+  - python --version
+  - python setup.py install
+  # - pip install -e .
   
   # do tests
   - python -c "import pykima"
