+++ .travis.yml
@@ -8,8 +8,8 @@ os:
 compiler:
   - gcc
 
-before_install:
-  - if [[ "$TRAVIS_OS_NAME" == "osx" ]]; then brew update; fi
+install:
+  # - if [[ "$TRAVIS_OS_NAME" == "osx" ]]; then brew update; fi
   # setup miniconda
   - if [[ "$TRAVIS_OS_NAME" == "osx" ]]; then
       wget http://repo.continuum.io/miniconda/Miniconda-latest-MacOSX-x86_64.sh -O ~/miniconda.sh;
@@ -18,19 +18,22 @@ before_install:
     fi
   - bash ~/miniconda.sh -b -p $HOME/miniconda
   - export PATH="$HOME/miniconda/bin:$PATH"
-  - conda update --yes conda
+  - hash -r
+  - conda config --set always_yes yes --set changeps1 no
+  - conda update -q conda
+  - conda info -a
   - python --version
   - pip --version
 
-install:
+script:
+  - conda create -q -n test-environment python=2.7 numpy
+  - source activate test-environment
   # - conda install --yes numpy scipy matplotlib pandas pytest
   # - make -j
   # now the pykima package
   - pwd
   - ls
   - python setup.py install
-
-script:
   # do tests
   - python -c "import pykima"
   - kima-showresults -h
