### 2018-01-03 513cdaf
install corner; almost final travis, maybe
