### 2018-01-03 c0b541f
add back osx and do slow tests
