### 2018-01-03 j-faria/kima#21 @j-faria
Travis now compiles the source code and the examples, and also builds the `pykima` package.
I added some tests to the package itself (the python code) and a new example that samples from the default priors and tests them. Also added builds in both Python 2.7 and 3.5 / 3.6.

My next goal is to add some coverage information, but in the meantime, I think we can close this issue. @barentsen, feel free to re-open if you disagree.
