### 2018-01-03 bfc40cd
cleanup; export c++ compiler in top level makefile
