+++ Makefile
@@ -1,15 +1,11 @@
-# Put the directory of 'DNest4/code' into this variable
 DNEST4_PATH = DNest4/code
 EIGEN_PATH = eigen
-# CELERITE_PATH = celerite/cpp/include
 
-includes = -I$(DNEST4_PATH) -I$(EIGEN_PATH) 
-# -I$(CELERITE_PATH)
+export CXX = g++
 
-CXX = g++
 CXXFLAGS = -pthread -std=c++11 -O3 -DNDEBUG -w -DEIGEN_MPL2_ONLY
 LIBS = -L$(DNEST4_PATH) -ldnest4 -L/usr/local/lib
-
+includes = -I$(DNEST4_PATH) -I$(EIGEN_PATH) 
 
 SRCDIR = ./src
 SRCS =\
