### 2018-01-03 eec00e4
simplify makefile; update DNest4
