### 2018-01-03 77d4e82
fix bug in showresults interactive; add coveralls to travis
