### 2018-01-03 a3c0aec
forgot pytest-cov in travis
