### 2018-01-03 239d6f0
clean; add a couple of tests
