+++ pykima/display.py
@@ -520,90 +520,6 @@ class KimaResults(object):
         plt.show()
 
 
-
-    def make_plot31(self, star_mass=1.0, trend=0., points=True):
-        from astropy import units as u
-
-        T = np.exp(self.T) if self.log_period else self.T
-        A, E = self.A, self.E
-
-        T = T * u.day
-        A = A * u.meter/u.second
-
-        Mmj = get_planet_mass(T.value, A.value, E, star_mass=star_mass, full_output=True)
-        Mmj = Mmj[2] * u.jupiterMass
-        # aliases1 = get_aliases(0.85359165)
-        # aliases2 = get_aliases(3.691)
-        # aliases3 = get_aliases(9.03580275887)
-
-        fig = plt.figure()
-
-        ax1 = fig.add_subplot(2,1,1)
-        #plot(truth[1009:1009 + int(truth[1008])]/log(10.), log10(truth[1018:1018 + int(truth[1008])]), 'ro', markersize=7)
-        #hold(True)
-        if points:
-            # ax1.loglog(T, Mmj, '.', markersize=1)
-            ax1.scatter(T, Mmj, c=E, s=5)
-            ax1.set_xscale('log')
-            ax1.set_yscale('log')
-        else:
-            ax1.hexbin(T[::100], A[::100], gridsize=50, bins='log', xscale='log', yscale='log',
-                       cmap=plt.get_cmap('afmhot_r'))
-        # data = np.vstack([np.exp(T[::100]), A[::100]]).T
-        # sns.jointplot(x=np.exp(T[::100]), y=A[::100], kind="hex", color="k");
-        # sns.kdeplot(data=np.exp(T[::100]), data2=A[::100], bw=[0.1, 1], shade=True, ax=ax1)
-
-        ax1.set_ylabel(r'Mass (Mjup)')
-
-        # mu = self.posterior_sample[:, self.index_component-1]
-        # ax = fig.add_subplot(2,2,2, sharey=ax1)
-        # ax.hist(mu, bins=30, orientation="horizontal")
-        # ax.set_ylim(ax1.get_ylim())
-        # ax.set_yscale('log')
-
-        ax2 = fig.add_subplot(2,1,2, sharex=ax1)
-        #plot(truth[1009:1009 + int(truth[1008])]/log(10.), truth[1038:1038 + int(truth[1008])], 'ro', markersize=7)
-        #hold(True)
-        if points:
-            ax2.semilogx(T, E, 'b.', markersize=2)
-        else:
-            ax2.hexbin(T[::100], E[::100], gridsize=50, bins='log', xscale='log',
-                       cmap=plt.get_cmap('afmhot_r'))
-        # ax.axvline(x=0.85359165, color='r')
-        # ax.axvline(x=3.691, color='r')
-        # ax.vlines([3.9359312722691815, 3.8939651463077287, 3.8528844910117557, 3.8126615742655399, 3.7732698100471724, 3.7346836998277957, 3.6968787775300003, 3.6598315577957377, 3.6235194873338745, 3.5879208991356037, 3.5530149693624242, 3.5187816767264146], ymin=0, ymax=1)
-        # ax.vlines(aliases1, ymin=0, ymax=1, color='g', alpha=0.4)
-        # ax.vlines(aliases2, ymin=0, ymax=1, color='y', alpha=0.4)
-        # ax.vlines(aliases3, ymin=0, ymax=1, color='c', alpha=0.4)
-        
-        ax2.set_xlim([10, 1e7])
-        ax2.set_xlabel(r'(Period/days)')
-        ax2.set_ylabel('Eccentricity')
-
-        for ax in (ax1,ax2):
[…]
