### 2018-01-04 j-faria/kima#20 @j-faria
This is now done, so I think I can close this issue.
