### 2018-01-04 j-faria/kima#23 @j-faria
I have used `dnest4`'s Python bindings, which do make it easier to specify the model, priors, etc. Well, easier just because they are specified in Python instead of C++, so the development might be a bit faster.
Sampling performance still suffers though, because the likelihood is then specified in Python (also I didn't find a way to parallelize the sampler into threads using the bindings).

Another issue is that the trans-dimensional sampling is not included in the Python bindings, as far as I can tell. So I still see kima as a C++ package, using the C++ code from DNest4, even if I completely agree that a Python-only package would simplify installation.
