### 2018-01-04 j-faria/kima#24 @j-faria
@barentsen, I agree with your suggestion to cite other packages. However, I'm not sure if this should be done in the paper. Is it common in JOSS papers to cite other _similar_ packages? I would be happy to rephrase a bit the text in the paper to highlight kima's key advantages, but I don't think a list of similar packages should be included in the paper.

Regarding the documentation, I added some words [here](https://github.com/j-faria/kima/wiki/1.-What-is-kima) and was thinking of adding something like a "see also" section on the main readme.
