### 2018-01-04 a99575a
add license badge to readme [skip ci]
