### 2018-01-07 e7b85f7
rephrase paper and add a couple of citations
