+++ paper/paper.md
@@ -32,38 +32,52 @@ bibliography: paper.bib
 
 # Summary
 
-The radial-velocity (RV) method is one of the most successful in the detection of exoplanets.
+The radial-velocity (RV) method is one of the most 
+successful in the detection of exoplanets.
 An orbiting planet induces a gravitational pull on its host star,
 which is observed as a periodic variation of the velocity of the star 
 in the direction of the line of sight.
-By measuring the associated wavelength shifts in stellar spectra, a RV timeseries is constructed.
-These data provide information about the presence of (one or more) planets
-and allow for the planet mass
-and several orbital parameters to be determined. <!--  [e.g. @Fischer2016]. -->
+By measuring the associated wavelength shifts in stellar spectra, 
+a RV timeseries is constructed.
+These data provide information about the presence of (one or more) 
+planets and allow for the planet mass(es) and 
+several orbital parameters to be determined
+[e.g. @Fischer2016].
 
 One of the main barriers to the detection of Earth-like planets with RVs
 is the intrinsic variations of the star,
 which can easily mimic or hide true RV signals of planets.
 Gaussian processes (GP) are now seen as a promising tool to model 
 the correlated noise that arises from stellar-induced RV variations. 
-<!-- [e.g. @Haywood2014]. -->
+[e.g. @Haywood2014].
 
 
-**kima** is a package for the detection and characterization of exoplanets using RV data.
+**kima** is a package for the detection and characterization of 
+exoplanets using RV data.
 It fits a sum of Keplerian curves to a timeseries of RV measurements, 
 using the Diffusive Nested Sampling algorithm [@Brewer2011]
 to sample from the posterior distribution of the model parameters. 
 This algorithm is extremely efficient on the 
 multimodal and correlated posteriors that often arise in this problem.
 
-Additionally, **kima** also calculates the fully marginalized likelihood 
-(or _evidence_) <!-- ; see @Kass1995) --> 
-either for a model with a fixed number $N_p$ of Keplerian signals,
+<!-- Additionally, -->
+Unlike similar open-source packages, 
+**kima** calculates the fully marginalized likelihood, or _evidence_,
+<!--; see @Kass1995)  -->
+both for a model with a fixed number $N_p$ of Keplerian signals,
 or marginalising over $N_p$.
-For this latter task we sample from the posterior distribution for $N_p$ itself,
+For this latter task, $N_p$ itself is a free parameter
+and we sample from its posterior distribution
 using the trans-dimensional method proposed by @Brewer2014.
 <!-- [see also @Brewer2015, @Faria2016] -->
 <!--  -->
+<!-- This means that **kima** can be used to compare models.  -->
+Because **kima** uses the Diffusive Nested Sampling algorithm,
+the evidence values are still accurate
+when the likelihood function contains phase changes
+which would make other algorithms (such as thermodynamic integration) unreliable. 
+<!--  -->
+
 Moreover, **kima** can use a GP with a quasi-periodic kernel
 as a noise model, to deal with activity-induced signals.
 The hyperparameters of the GP are inferred together with the orbital parameters.
@@ -71,8 +85,8 @@ Priors for each of the parameters can be easily set by the user,
 with a broad choice of standard probability distributions already implemented.
 
 
-The code is written in C\texttt{++}, and includes several Python scripts
-that can be used for the analysis of the results.
+The code is written in C\texttt{++}, but also includes a helper Python package,
+`pykima`, which facilitates the analysis of the results.
[…]
