+++ paper/paper.bib
@@ -41,7 +41,7 @@
   number = {4},
   urldate = {2015-12-03},
   journal = {Statistics and Computing},
-  author = {Brewer, B. J. and P{\'a}rtay, Livia B. and Cs{\'a}nyi, G{\'a}bor},
+  author = {Brewer, B. J. and P{\'a}rtay, L. B. and Cs{\'a}nyi, G.},
   month = oct,
   year = {2011},
   pages = {649--656},
@@ -62,4 +62,40 @@
   year = {2016},
   pages = {A31},
   file = {/home/joao/phd/papers/Faria_etal_2016.pdf}
+}
+
+
+@article{Haywood2014,
+  title = {Planets and Stellar Activity: Hide and Seek in the {{CoRoT}}-7 System},
+  volume = {443},
+  issn = {0035-8711},
+  shorttitle = {Planets and Stellar Activity},
+  doi = {10.1093/mnras/stu1320},
+  timestamp = {2015-12-03T18:36:39Z},
+  urldate = {2015-01-08},
+  journal = {Monthly Notices of the Royal Astronomical Society},
+  author = {Haywood, R. D. and Collier Cameron, A. and Queloz, D. and Barros, S. C. C. and Deleuil, M. and Fares, R. and Gillon, M. and Lanza, A. F. and Lovis, C. and Moutou, C. and Pepe, F. and Pollacco, D. and Santerne, A. and S{\'e}gransan, D. and Unruh, Y. C.},
+  month = sep,
+  year = {2014},
+  keywords = {planetary systems,stars: activity,stars: individual: CoRoT-7,techniques: radial velocities},
+  pages = {2517--2531},
+  file = {/home/joao/phd/papers/Haywood_etal_2014.pdf;/home/joao/phd/papers/Haywood_etal_2014.pdf;/home/joao/phd/papers/Haywood_etal_2014.pdf},
+  groups = {CoRoT-7}
+}
+
+
+@article{Fischer2016,
+  title = {State of the {{Field}}: {{Extreme Precision Radial Velocities}}},
+  volume = {128},
+  issn = {1538-3873},
+  shorttitle = {State of the {{Field}}},
+  doi = {10.1088/1538-3873/128/964/066001},
+  language = {en},
+  timestamp = {2017-07-13T14:27:06Z},
+  number = {964},
+  journal = {Publications of the Astronomical Society of the Pacific},
[…]
