### 2018-01-07 #487 @j-faria
Hi @barentsen,
I think I have now addressed all the issues you raised in your first review.

 - The C++ code, together with the examples, is now compiled in Travis automatically (and the related issues were fixed)
 - I created the `pykima` package as suggested, which now makes available a `kima-showresults` script and can also be imported from a Python interpreter. I'm not sure if it makes sense to add this package to pypi, so that it is `pip`-installable, but would be happy to do so.
 - Added some automated tests to Travis. These test the Python package and also compile and run a new example to check the default priors.
 - Added some words to the wiki with advantages and disadvantages of kima when compared with similar packages. Also rephrased a bit the paper to highlight kima's features. But I don't think a full list of similar packages should be in the paper.
