### 2018-01-07 #487 @barentsen
I hope to continue the review soon, but unfortunately I'll be tied up with the AAS meeting during all of next week.
