### 2018-01-13 694ac07
use (copy of) dnest4.classic; this removes pandas dependence; cleanup; should fix #9
