+++ pykima/dnest4.py
@@ -1,301 +1,0 @@
-import copy
-import numpy as np
-import matplotlib.pyplot as plt
-
-from time import time
-import pandas as pd
-
-try:
-    from tqdm import tqdm
-except ImportError:
-    def tqdm(x, *args):
-        return x
-
-def logsumexp(values):
-    biggest = np.max(values)
-    x = values - biggest
-    result = np.log(np.sum(np.exp(x))) + biggest
-    return result
-
-def logdiffexp(x1, x2):
-    biggest = x1
-    xx1 = x1 - biggest
-    xx2 = x2 - biggest
-    result = np.log(np.exp(xx1) - np.exp(xx2)) + biggest
-    return result
-
-def loadfiles():
-    k = {'delim_whitespace':True, 'header':None, 'comment':'#'}
-    t1 = time()
-    levels_orig = pd.read_csv("levels.txt", **k).values
-    sample_info = pd.read_csv("sample_info.txt", **k).values
-    sample = pd.read_csv("sample.txt", **k).values
-    t2 = time()
-    took = t2 - t1
-    return levels_orig, sample_info, sample, took
-
-def plot_levels_vs_iteration(sample_info):
-    plt.figure()
-    plt.plot(sample_info[:,0])
-    plt.xlabel("Iteration")
-    plt.ylabel("Level")
-
-def plot_compression_acceptance_vs_levels(levels_orig):
-    plt.figure()
-    plt.subplot(2,1,1)
-    plt.plot(np.diff(levels_orig[:,0]))
-    plt.ylabel("Compression")
-    plt.xlabel("Level")
-    xlim = plt.gca().get_xlim()
-    plt.axhline(-1., color='r')
-    plt.axhline(-np.log(10.), color='g')
-    plt.ylim(ymax=0.05)
-
-    plt.subplot(2,1,2)
-    good = np.nonzero(levels_orig[:,4] > 0)[0]
-    plt.plot(levels_orig[good,3]/levels_orig[good,4])
-    plt.xlim(xlim)
-    plt.ylim([0., 1.])
-    plt.ylabel("MH Acceptance")
-    plt.xlabel("Level")
-
-
-def plot_logL_weights_vs_logX(z, sample_info, levels, numResampleLogX,
-                              logx_samples, logz_estimates, P_samples, zoom_in):
-    plt.figure()
-
-    plt.subplot(2,1,1)
-    plt.plot(logx_samples[:,z], sample_info[:,1], 'b.', label='Samples')
-    plt.plot(levels[1:,0], levels[1:,1], 'r.', label='Levels')
-    plt.legend(numpoints=1, loc='lower left')
-    plt.ylabel('log(L)')
-    plt.title(str(z+1) + "/" + str(numResampleLogX) + ", log(Z) = " + str(logz_estimates[z][0]))
-    # Use all plotted logl values to set ylim
-    combined_logl = np.hstack([sample_info[:,1], levels[1:, 1]])
-    combined_logl = np.sort(combined_logl)
-    lower = combined_logl[int(0.1*combined_logl.size)]
-    upper = combined_logl[-1]
-    diff = upper - lower
-    lower -= 0.05*diff
-    upper += 0.05*diff
-    if zoom_in:
-        plt.ylim([lower, upper])
-
-    xlim = plt.gca().get_xlim()
-
-    plt.subplot(2,1,2)
-    plt.plot(logx_samples[:,z], P_samples[:,z], 'b.')
-    plt.ylabel('Posterior Weights')
-    plt.xlabel('log(X)')
-    plt.xlim(xlim)
-
-
-def postprocess(temperature=1., numResampleLogX=1, plot=True, loaded=[],
-                cut=0, save=True, zoom_in=True, compression_bias_min=1,
[…]
