### 2018-01-14 967b4fd
remove dependence on pandas
