### 2018-01-14 c84ee6e
delete old scripts folder
