+++ scripts/dnest4.py
@@ -1,279 +1,0 @@
-import copy
-import numpy as np
-import matplotlib.pyplot as plt
-
-from time import time
-import pandas as pd
-
-try:
-    from tqdmmm import tqdm
-except ImportError:
-    def tqdm(x, *args):
-        return x
-
-def logsumexp(values):
-    biggest = np.max(values)
-    x = values - biggest
-    result = np.log(np.sum(np.exp(x))) + biggest
-    return result
-
-def logdiffexp(x1, x2):
-    biggest = x1
-    xx1 = x1 - biggest
-    xx2 = x2 - biggest
-    result = np.log(np.exp(xx1) - np.exp(xx2)) + biggest
-    return result
-
-
-def postprocess(temperature=1., numResampleLogX=1, plot=True, loaded=[],
-                cut=0., save=True, zoom_in=True, compression_bias_min=1,
-                compression_scatter=0., moreSamples=1., compression_assert=None):
-    if len(loaded) == 0:
-        # t1 = time()
-        # levels_orig1 = np.atleast_2d(np.loadtxt("levels.txt"))
-        # sample_info1 = np.atleast_2d(np.loadtxt("sample_info.txt"))
-        # sample1 = np.atleast_2d(np.loadtxt("sample.txt"))
-        # print 'Took %f sec' % (time() - t1, )
-
-        k = {'delim_whitespace':True, 'header':None, 'comment':'#'}
-        t1 = time()
-        levels_orig = pd.read_csv("levels.txt", **k).values
-        sample_info = pd.read_csv("sample_info.txt", **k).values
-        sample = pd.read_csv("sample.txt", **k).values
-        print('Took %f sec to read files' % (time() - t1, ))
-
-        # assert np.allclose(sample1, sample)
-        # assert np.allclose(sample_info1, sample_info)
-        # assert np.allclose(levels_orig1, levels_orig)
-
-        # I believe this is the fastest way to get the 2nd line 
-        # and doesn't load the whole file into memory
-        fp = open('sample.txt')
-        for i, line in enumerate(fp):
-            if i==1: 
-                sample_file_header = line
-            elif i>2: 
-                break
-        fp.close()
-
-        #if(sample.shape[0] == 1):
-        #   sample = sample.T
-    else:
-        levels_orig, sample_info, sample = loaded[0], loaded[1], loaded[2]
-
-    # Remove regularisation from levels_orig if we asked for it
-    if compression_assert is not None:
-        levels_orig[1:,0] = -np.cumsum(compression_assert*np.ones(levels_orig.shape[0] - 1))
-
-    sample = sample[int(cut*sample.shape[0]):, :]
-    sample_info = sample_info[int(cut*sample_info.shape[0]):, :]
-
-    if sample.shape[0] != sample_info.shape[0]:
-        print('# Size mismatch. Truncating...')
-        lowest = np.min([sample.shape[0], sample_info.shape[0]])
-        sample = sample[0:lowest, :]
-        sample_info = sample_info[0:lowest, :]
-
-    if plot:
-
-        plt.figure()
-        plt.plot(sample_info[:,0])
-        plt.xlabel("Iteration")
-        plt.ylabel("Level")
-
-        plt.figure()
-        plt.subplot(2,1,1)
-        plt.plot(np.diff(levels_orig[:,0]))
-        plt.ylabel("Compression")
-        plt.xlabel("Level")
-        xlim = plt.gca().get_xlim()
-        plt.axhline(-1., color='r')
-        plt.axhline(-np.log(10.), color='g')
-        plt.ylim(ymax=0.05)
-
-        plt.subplot(2,1,2)
-        good = np.nonzero(levels_orig[:,4] > 0)[0]
[…]
