+++ scripts/joss_paper_figure.py
@@ -1,156 +1,0 @@
-import os, sys
-import cPickle
-import numpy as np
-import matplotlib.pyplot as plt
-import matplotlib.gridspec as gridspec
-import matplotlib.image as image
-# import tqdm
-
-### use mpl 1.0 style
-import matplotlib as mpl
-mpl.style.use('classic')
-
-
-import seaborn.apionly as sns
-colors = sns.color_palette("colorblind")
-sns.set_palette(sns.color_palette("colorblind"))
-
-from keplerian import keplerian
-from display import DisplayResults
-# sys.path.append('/home/joao/Work/OPEN')
-# from OPEN.ext.keplerian import keplerian
-
-
-sys.path.append('.')
-from styler import styler
-
-# res = DisplayResults('')
-# res.sample = np.atleast_2d(np.loadtxt('sample.txt'))
-
-res = cPickle.load(open('BL2009_joss_figure1.pickle'))
-
-
-@styler
-def f(fig, *args, **kwargs):
-
-	gs = gridspec.GridSpec(3, 6)
-	# gs.update(hspace=0.4, wspace=0.3, bottom=0.08, top=0.97, left=0.08, right=0.98)
-	ax1 = plt.subplot(gs[:2, :4])
-	ax2 = plt.subplot(gs[1, 4:6])
-	ax3 = plt.subplot(gs[2, 0:2])
-	ax4 = plt.subplot(gs[2, 2:4])
-	ax5 = plt.subplot(gs[2, 4:6])
-
-
-
-	samples = res.get_sorted_planet_samples()
-	samples, mask = \
-	    res.apply_cuts_period(samples, 90, None, return_mask=True)
-
-
-	t, y, yerr = res.data.T
-	over = 0.1
-	tt = np.linspace(t[0]-over*t.ptp(), t[-1]+over*t.ptp(), 
-	                 10000+int(100*over))
-
-	y = res.data[:,1].copy()
-	yerr = res.data[:,2].copy()
-
-	# select random `ncurves` indices 
-	# from the (sorted, period-cut) posterior samples
-	ncurves = 10
-	ii = np.random.randint(samples.shape[0], size=ncurves)
-
-	## plot the Keplerian curves
-	for i in ii:
-	    v = np.zeros_like(tt)
-	    pars = samples[i, :].copy()
-	    nplanets = pars.size / res.n_dimensions
-	    for j in range(nplanets):
-	        P = pars[j + 0*res.max_components]
-	        K = pars[j + 1*res.max_components]
-	        phi = pars[j + 2*res.max_components]
-	        t0 = t[0] - (P*phi)/(2.*np.pi)
-	        ecc = pars[j + 3*res.max_components]
-	        w = pars[j + 4*res.max_components]
-	        v += keplerian(tt, P, K, ecc, w, t0, 0.)
-
-	    vsys = res.posterior_sample[mask][i, -1]
-	    v += vsys
-	    ax1.plot(tt, v, alpha=0.2, color='k')
-
-	ax1.errorbar(*res.data.T, fmt='o', mec='none', capsize=0, ms=4,
-		         color=sns.color_palette()[2])
-	ax1.set(xlabel='Time [days]', ylabel='RV [m/s]',)
-
-
-
-
-	n, bins, _ = ax2.hist(res.posterior_sample[:, res.index_component], 
-		                  bins=np.arange(2, 4)-0.5, align='mid', rwidth=0.5, 
-		                  color='k', alpha=0.5)
[…]

[…]
