### 2018-01-14 f9cd4ba
remove old plots of prior samples, there is an example now
