### 2018-01-14 bf404da
update examples readme
