+++ examples/README.md
@@ -1,7 +1,8 @@
 This folder contains examples of how to run **kima**.  
 
-Note that this is also the test suite for the package. 
-When you type `make test` in the kima directory, all the examples are executed with a given random seed and the output is checked against a previous run.
+This is also the test suite for the package. 
+When you type `make test` in the kima directory, all the examples are executed with a given random seed 
+and the output is checked against a previous run. **that's the goal, not yet happening!**
 
 #### Use the examples
 
@@ -12,11 +13,11 @@ make
 ./run
 ```
 
-But in order to keep things organized, it is better to copy one of these folders into another place.
+But in order to keep things organized, it is better to copy one of these folders to another place.
 In this way you can use the example as a template and modify it to your will.  
 You will need to change:
 
 1. the `Makefile`; change the `KIMA_DIR` definition to point to the kima directory
    
-After this, you should be able to `./run` the copied example from a new directory location.
+After this, you should be able to `make` and `./run` the copied example from the new directory location.
 
