### 2018-01-14 18469e4
remove python checks from makefile
