+++ Makefile
@@ -34,38 +34,10 @@ examples: $(DNEST4_PATH)/libdnest4.a $(OBJS)
 	@make -C examples/CoRoT7
 	@make -C examples/many_planets
 	@make -C examples/default_priors
-	
+
 
 $(DNEST4_PATH)/libdnest4.a:
 	make -C $(DNEST4_PATH)
-
-pythoncheck:
-	$(eval NOPACK=0)
-	@python -c "import numpy" 2>/dev/null ; \
-		if [ $$? -eq 1 ] ; \
-		then $(eval $NOPACK=1) echo "numpy does not seem to be installed!"; fi
-
-	@python -c "import scipy" 2>/dev/null ; \
-		if [ $$? -eq 1 ] ; \
-		then $(eval $NOPACK=1) echo "scipy does not seem to be installed!"; fi
-
-	@python -c "import pandas" 2>/dev/null ; \
-		if [ $$? -eq 1 ] ; \
-		then $(eval $NOPACK=1) echo "pandas does not seem to be installed!"; fi
-
-	@python -c "import matplotlib" 2>/dev/null ; \
-		if [ $$? -eq 1 ] ; \
-		then $(eval $NOPACK=1) echo "matplotlib does not seem to be installed!"; fi
-
-	@python -c "import corner" 2>/dev/null ; \
-		if [ $$? -eq 1 ] ; \
-		then $(eval $NOPACK=1) echo "corner does not seem to be installed!"; fi
-
-	@if [ $(NOPACK) -eq 1 ] ; \
-		then echo \
-			"Some required Python packages are missing," \
-			"kima will still work but not the analysis of results.";\
-		fi
 
 
 clean:
