### 2018-01-14 f32321e
import from classic in __init__
