+++ pykima/__init__.py
@@ -1,3 +1,3 @@
-from .dnest4 import postprocess
+from .classic import postprocess
 from .display import KimaResults
 from .showresults import showresults
