### 2018-01-14 41aaa8b
remove pythoncheck
