### 2018-01-17 2576fe0
add the checkpriors script
