+++ setup.py
@@ -17,6 +17,7 @@ setup(name='pykima',
       entry_points={
         'console_scripts': [
             'kima-showresults = pykima.showresults:showresults',
+            'kima-checkpriors = pykima.check_priors:main',
             ]
         },
      )
