### 2018-01-18 #487 @j-faria
Hi @barentsen! Sorry to put pressure, but do you think you will have some time to continue the review in the next few days? I'll have my PhD defense in the end of next week, and it would be awesome to "announce" kima there for the first time :)
