### 2018-02-02 d07fa82
revert to DNest4
