+++ src/RVConditionalPrior.h
@@ -25,8 +25,8 @@ class RVConditionalPrior:public DNest4::ConditionalPrior
 		void from_prior(DNest4::RNG& rng);
 
 		double log_pdf(const std::vector<double>& vec) const;
-		void from_uniform(std::vector<double>& vec, int component_id) const;
-		void to_uniform(std::vector<double>& vec, int component_id) const;
+		void from_uniform(std::vector<double>& vec) const;
+		void to_uniform(std::vector<double>& vec) const;
 
 		void print(std::ostream& out) const;
 		static const int weight_parameter = 1;
