+++ Makefile
@@ -1,3 +1,5 @@
+VERBOSE ?= 0
+
 DNEST4_PATH = DNest4/code
 EIGEN_PATH = eigen
 
@@ -21,23 +23,48 @@ HEADERS=$(subst .cpp,.h,$(SRCS))
 all: main examples
 
 %.o: %.cpp
+ifeq ($(VERBOSE), 0)
+	@echo "\033[0;33m Compiling:\033[0m" $<
+	@$(CXX) -c $(includes) -o $@ $< $(CXXFLAGS)
+else
 	$(CXX) -c $(includes) -o $@ $< $(CXXFLAGS)
+endif
 
 
 main: $(DNEST4_PATH)/libdnest4.a $(OBJS)
-	$(CXX) -o main $(OBJS) $(LIBS) $(CXXFLAGS)
+ifeq ($(VERBOSE), 0)
+	@echo "\033[0;33m Linking\033[0m "
+	@$(CXX) -o kima $(OBJS) $(LIBS) $(CXXFLAGS)
+else
+	$(CXX) -o kima $(OBJS) $(LIBS) $(CXXFLAGS)
+endif
 
 
 .PHONY: examples
 examples: $(DNEST4_PATH)/libdnest4.a $(OBJS)
+ifeq ($(VERBOSE), 0)
+	@make -s -C examples/BL2009
+	@echo "\033[0;33m Compiling example\033[0m BL2009"
+	@make -s -C examples/CoRoT7
+	@echo "\033[0;33m Compiling example\033[0m CoRoT7"
+	@make -s -C examples/many_planets
+	@echo "\033[0;33m Compiling example\033[0m many_planets"
+	@make -s -C examples/default_priors
+	@echo "\033[0;33m Compiling example\033[0m default_priors"
+else
 	@make -C examples/BL2009
 	@make -C examples/CoRoT7
 	@make -C examples/many_planets
 	@make -C examples/default_priors
-
+endif
 
 $(DNEST4_PATH)/libdnest4.a:
-	make -C $(DNEST4_PATH)
+ifeq ($(VERBOSE), 0)
+	@echo "\033[0;33m Compiling \033[0m DNest4"
+	@make -s -C $(DNEST4_PATH) libdnest4.a
+else
+	make -C $(DNEST4_PATH) libdnest4.a
+endif
 
 
 clean:
