### 2018-02-05 0a23ac4
Jeffreys -> LogUniform
