+++ src/default_planet_priors.h
@@ -3,8 +3,8 @@
 
 #include "DNest4.h"
 
-Jeffreys *Pprior = new Jeffreys(1.0, 1E5); // days
-ModifiedJeffreys *Kprior = new ModifiedJeffreys(1.0, 2E3); // m/s
+LogUniform *Pprior = new LogUniform(1.0, 1E5); // days
+ModifiedLogUniform *Kprior = new ModifiedLogUniform(1.0, 2E3); // m/s
 
 Uniform *eprior = new Uniform(0., 1.);
 Uniform *phiprior = new Uniform(0.0, 2*M_PI);
