+++ examples/BL2009/kima_setup.cpp
@@ -28,10 +28,10 @@ RVmodel::RVmodel()
 {
     // priors as in Balan & Lahav (2009, DOI: 10.1111/j.1365-2966.2008.14385.x)
     Cprior = new Uniform(-2000, 2000);
-    Jprior = new ModifiedJeffreys(1.0, 2000.); // additional white noise, m/s
+    Jprior = new ModifiedLogUniform(1.0, 2000.); // additional white noise, m/s
 
-    Pprior = new Jeffreys(0.2, 15E3); // days
-    Kprior = new ModifiedJeffreys(1.0, 2E3); // m/s
+    Pprior = new LogUniform(0.2, 15E3); // days
+    Kprior = new ModifiedLogUniform(1.0, 2E3); // m/s
 
     eprior = new Uniform(0., 1.);
     phiprior = new Uniform(0.0, 2*M_PI);
