### 2018-02-14 7929f52
add version number; more info and license for pykima
