### 2018-02-14 b635c58
use a prior for the fiber offset, still the uniform though
