### 2018-02-14 c895b38
bug fix when printing sample.txt header
