### 2018-02-16 #487 @barentsen
I apologize for the delay.  I expect to have time to continue the review on Tuesday.
