### 2018-02-22 ba17db7
allow for comparison samples to be plotted with kima-checkpriors
