### 2018-02-22 605b46d
fix order of fiber offset and trend; show trend in plot; gp parameter units
