### 2018-02-22 295a2d8
test suite for some of the examples
