+++ examples/README.md
@@ -1,8 +1,10 @@
 This folder contains examples of how to run **kima**.  
 
 This is also the test suite for the package. 
-When you type `make test` in the kima directory, all the examples are executed with a given random seed 
-and the output is checked against a previous run. **that's the goal, not yet happening!**
+When you type `make test` in the kima directory, 
+all the examples are executed with a given random seed 
+and the output is checked against a previous run. 
+**that's the goal, many_planets still has problems!**
 
 #### Use the examples
 
@@ -12,6 +14,8 @@ The easiest way to use these examples is to go in a given directory and type
 make
 ./run
 ```
+
+*note:* replace `make` with `CXX=compiler make` if you use a different compiler
 
 But in order to keep things organized, it is better to copy one of these folders to another place.
 In this way you can use the example as a template and modify it to your will.  
