### 2018-02-26 b865dbb
helper methods for data; info on m/s transformation
