+++ src/Data.h
@@ -5,6 +5,8 @@
 #include <algorithm>
 #include <cmath>
 
+#include "DNest4.h"
+
 class Data
 {
 	private:
@@ -18,14 +20,18 @@ class Data
 
 		// Getters
 		int N() const {return t.size();}
+
 		const std::vector<double>& get_t() const { return t; }
 		double get_t_min() const { return *std::min_element(t.begin(), t.end()); }
 		double get_t_max() const { return *std::max_element(t.begin(), t.end()); }
+		double get_t_middle() const { return get_t_min() + 0.5*(get_t_max() - get_t_min()); }
 
 		const std::vector<double>& get_y() const { return y; }
 		double get_y_min() const { return *std::min_element(y.begin(), y.end()); }
 		double get_y_max() const { return *std::max_element(y.begin(), y.end()); }
-		double get_RV_span() const {return get_y_max() - get_y_min();}
+		double get_RV_span() const { return get_y_max() - get_y_min(); }
+		double get_RV_var() const { return DNest4::variance(y); }
+		double get_RV_std() const { return std::sqrt(DNest4::variance(y)); }
 		
 		const std::vector<double>& get_sig() const { return sig; }
 		double topslope() const {return std::abs(get_y_max() - get_y_min()) / (t.back() - t.front());}
