### 2018-02-26 0624352
change variance to be method of Data
