### 2018-02-27 13d95d7
Delete LICENSE.md
