### 2018-03-01 143a0bd
plot 3 does nothing without planets; bug in printing info
