### 2018-03-02 e9913ba
change prior for the fiber offset
