### 2018-03-02 5652a8a
auto data and clean ups
