### 2018-03-02 2bbbc8d
for braces; reference time for slope is middle of times
