### 2018-03-02 c15a1fb
nasty bug, fixed and simplified ifs
