### 2018-03-02 cfb5313
bookkeeping variables
