### 2018-03-15 2bf6947
actually install the package; fixes #33; python 3 stuff in check_priors
