### 2018-03-15 3b9cdb9
in Travis: install not develop; import from different directory
