### 2018-03-30 #487 @j-faria
Hi @arfon, I'm not sure; I think Daniel Foreman-Mackey has reviewed JOSS papers before, and Jake VanderPlas also. They're both working in related areas (read astronomy).
Otherwise, is there a list of joss reviewers?
