### 2018-04-20 #487 @j-faria
Hi @nespinoza, thank you for accepting to review kima. Please feel free to comment on the repository and let me know if I can answer any questions that might come up.
