### 2018-04-28 3b3dad4
new options for percentiles; cleanup; py3 issue with int
