### 2018-04-28 06860d5
add method to read multiple files or one file with multiple instruments
