+++ src/Data.h
@@ -9,11 +9,17 @@ class Data
 {
 	private:
 		std::vector<double> t, y, sig;
+		std::vector<int> obsi;
 
 	public:
 		Data();
-		//void load(const char* filename);
+		// to read data from one file, one instrument
 		void load(const char* filename, const char* units, int skip=2);
+		// to read data from one file, more than one instrument
+		void load_multi(const char* filename, const char* units, int skip=2);
+		// to read data from more than one file, more than one instrument
+		void load_multi(std::vector<char*> filenames, const char* units, int skip=2);
+
 		int index_fibers;
 
 		const char* datafile;
@@ -37,6 +43,7 @@ class Data
 		double get_RV_std() const { return std::sqrt(get_RV_var()); }
 		
 		const std::vector<double>& get_sig() const { return sig; }
+		const std::vector<int>& get_obsi() const { return obsi; }
 		double topslope() const {return std::abs(get_y_max() - get_y_min()) / (t.back() - t.front());}
 
 	// Singleton
