### 2018-04-29 386febf
count number of instruments
