### 2018-04-29 17b17ba
implement instrument offsets; common prior for all (to be set by the user)
