### 2018-04-29 0bc9939
new variables need to be defined
