+++ src/default_extra_priors.h
@@ -7,3 +7,9 @@ ModifiedLogUniform *Jprior = new ModifiedLogUniform(1.0, 99.); // additional whi
 // the prior probably should depend on the data,
 // -> a good default would be Uniform( -data.topslope(), data.topslope() )
 Uniform *slope_prior = new Uniform(-10, 10);
+
+
+// this default prior for the instrument offsets is awful!
+// the prior probably should depend on the data,
+// -> a good default would be Uniform( -data.get_RV_span(), data.get_RV_span() )
+Uniform *offsets_prior = new Uniform(-10, 10);
