### 2018-04-29 84a4bc1
separate some utility functions
