### 2018-05-03 95bf54a
save information from the data file
