### 2018-05-03 4bee497
new method to save the current model setup
