### 2018-05-03 f9d2354
trying to simplify the model checks
