### 2018-05-04 8805eb9
save data file info with multiple instruments
