### 2018-05-04 a2904fb
fix bug with "theoretical" samples
