### 2018-05-15 82a21fc
save_setup in all examples; small cleanup
