+++ examples/BL2009/kima_setup.cpp
@@ -9,9 +9,7 @@ using namespace DNest4;
 /* priors */
 //  data-dependent priors should be defined in the RVmodel() 
 //  constructor and use Data::get_instance() 
-
 #include "default_priors.h"
-
 
 
 const bool obs_after_HARPS_fibers = false;
@@ -36,6 +34,8 @@ RVmodel::RVmodel()
     eprior = new Uniform(0., 1.);
     phiprior = new Uniform(0.0, 2*M_PI);
     wprior = new Uniform(0.0, 2*M_PI);
+
+    save_setup();
 }
 
 
