### 2018-05-16 596668f
helper function for stellar mass in latex
