### 2018-05-31 #487 @nespinoza
Dear @j-faria,

 First of all, I apologize for taking so long in submitting my review; a lot has happened in the past few months and finding time to throughly go through the code and documentation was very hard. However, after running the examples, going through the documentation and most of the code, I can now (at last!) write my review.

 First of all, congratulations on putting this software together! It is definitely work that is going to be very useful for many people working on the area of planet detection and characterization. It is also very timely given the advent of new radial-velocity follow-up machines.

 As the authors can see, I have marked most of the review points above as I think they comply with the standards of a JOSS publication. I think the documentation is very good, and details quite nicely what the code does and how it does it. However, although the code runs smoothly, I found the examples to be not as thoroughly explained as I would like, thinking as a new user that wants to use the code for the things the code is supposed to do. In what follows, I will outline why I haven't marked the points above and what I would like to see in order to mark them as "done":

1. **Functionality and Performance**. There are two important claims in the paper that I think any user would like to see in action. The first --- and easiest to showcase --- is that _kima_ (a) fits Keplerian models to data, and the second is that (b) the algorithm used by _kima_, Diffusive Nested Sampling, is "extremely efficient" at sampling the kind of posteriors that appear in the problem of fitting multiple-planets to a radial-velocity dataset. Although both are implicitly showcased by the code --- after inspecting the posterior samples and comparing them to some expected results --- I think these should be explicitly showcased by the software to any user that is thinking on or is going to use it.

- [ ] For (a) I believe one of the outputs of _kima_ should be a set of posterior realizations of the model over the radial-velocity measurements, i.e., the same plot it is shown on the _kima_ repository (https://github.com/j-faria/kima).
- [ ] For point (b), I must say it is hard to actually _confirm_ this is the case for this particular case. For the purposes of the software, perhaps it would suffice to find a real-world example (or any of the examples that are already within the documentation), try to fit the dataset with some available software (e.g., any of the ones you mention in Section 1 of _kima_'s documentation), and show that _kima_ actually is more (or as) efficient and in what way. Alternatively, explaining why or in what sense Diffusive Nested Sampling is more efficient for this particular problem than other current solutions and the respective citation would be enough as well.

2. **Example usage** I really liked the examples within the package because they are either previously generated datasets from other authors or real-life problems. However, I'm inclined to think that they are not very useful by themselves to a new user wanting to use _kima_ for the first time. In particular (a) it is not explained what exactly are the inputs to _kima_ and how exactly they should be feeded to _kima_ (although it is obvious after checking the code that the inputs are times and radial-velocities, this might not be obvious to any user; e.g., some users might want to use some activity index instead of time as the independant variable in the GP, or add a linear trend with some activity index and so on. Also, does any format for times and radial-velocity work? Comma separated? Tabs? Note that _examples should be applications of these instructions, and not instructions on how to use the codes themselves_) and (b) although there are examples for setting different priors, there is no step-by-step example on how to setup a problem from scratch.

- [ ] For (a) and (b), I would strongly suggest the authors take one of the examples they provide and make a step-by-step resolution of the problem using _kima_, in a similar fashion as, e.g., this one for radvel: http://radvel.readthedocs.io/en/latest/quickstartcli.html#example-fit.

3. **Community guidelines** I think most of the points are not explicitly stated in the wiki. So please:
- [ ] Provide clear guidelines for how to contribute to the software.
- [ ] How to report issues or problems with the software.
- [ ] Who should users look for support.

I have not marked the **Version** review point, as the repository still doesn't have a release (I suppose because it is awaiting for review).

Best,

Néstor
