### 2018-06-04 ec1b3fe
deal with lack of kima_model_setup.txt (not very gracefully, fix #35); minor docs
