### 2018-06-05 be87330
Create issue templates
