+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -1,0 +1,11 @@
+---
+name: Feature request
+about: What should **kima** do that it still doesn't?
+
+---
+
+**Is your feature request related to a problem and/or open issue?**
+Please, provide a clear description of what the problem is or link back to the issue for context.
+
+**Describe the solution you'd like**
+If you know what a solution could look like, let us know.
