### 2018-06-06 6cbfd10
more python2/3 compatibility problems
