+++ pykima/utils.py
@@ -15,9 +15,9 @@ skip: 0
 def need_model_setup():
     print()
     print("[FATAL] Couldn't find the file kima_model_setup.txt")
-    print("Probably didn't include a call to save_setup()", end=' ')
+    print("Probably didn't include a call to save_setup()")
     print("in the RVModel constructor (this is the recommended solution).")
-    print("As a workaround, create a file called `kima_model_setup.txt`,", end=' ')
+    print("As a workaround, create a file called `kima_model_setup.txt`,")
     print("and add to it (after editting!) the following options:")
 
     print(template_setup)
