### 2018-06-06 042024f
better name for w
