### 2018-06-06 #487 @j-faria
Hi @nespinoza, thank you very much for the thorough review, I really appreciate it.

I agree completely with the three main points you raised as still missing. In the last couple of days I started addressing point 3, by creating issue and pull request templates, adding a `CONTRIBUTING.md` to the repo and changing the documentation. I will try to unify more clearly all these guidelines so that it becomes easier for a new user to contribute to *kima*.

To address point 2, I was planning to create a tutorial / getting started guide as you suggested. This is (in some form) already included in the wiki but the points you identified are indeed missing.

Point 1 is harder to address completely. Actually, there is already the possibility to plot a set of posterior realizations of the model over the RV measurements. It is currently undocumented though, so it's as good as not being there. But I will solve that very soon.
Regarding performance it will be challenging to choose a dataset that showcases well enough the capabilities of DNest4. Another possibility is to change our claims in the paper: currently *kima* stands apart from other packages because it can calculate the evidence for a 0, 1, 2, ... planet model *and* because it can sample the posterior for the number of planets Np. These two features are unique (to my knowledge). So those claims are more easily justified.

I will work on a complete answer to the points raised in your review in the upcoming two weeks.

Thank you again!
