### 2018-06-07 3d7ce68
change test to check FloatingPointError is raised
