### 2018-06-11 02f5519
show all plots at the end; fixes #41
