### 2018-06-11 a2955a8
improve some of the plots; fixes #39 and #40
