+++ pykima/display.py
@@ -399,16 +399,23 @@ class KimaResults(object):
 
     def make_plot1(self):
         """ Plot the histogram of the posterior for Np """
-        plt.figure()
-        n, bins, _ = plt.hist(self.posterior_sample[:, self.index_component], 100)
-        plt.xlabel('Number of Planets')
-        plt.ylabel('Number of Posterior Samples')
-        plt.xlim([-0.5, self.max_components+.5])
+        _, ax = plt.subplots(1,1)
+        # n, _, _ = plt.hist(self.posterior_sample[:, self.index_component], 100)
+        
+        bins = np.arange(self.max_components+2)
+        nplanets = self.posterior_sample[:, self.index_component]
+        n, _ = np.histogram(nplanets, bins=bins)
+        ax.bar(bins[:-1], n)
+
+        ax.set(xlabel='Number of Planets',
+               ylabel='Number of Posterior Samples',
+               xlim=[-0.5, self.max_components+.5],
+               xticks=np.arange(self.max_components+1),
+               title='Posterior distribution for $N_p$'
+              )
 
         nn = n[np.nonzero(n)]
-        print('probability ratios: ', nn.flat[1:] / nn.flat[:-1])
-
-        plt.show()
+        print('Np probability ratios: ', nn.flat[1:] / nn.flat[:-1])
 
 
     def make_plot2(self, bins=None):
@@ -427,27 +434,29 @@ class KimaResults(object):
         else:
             T = self.T
         
-        plt.figure()
+        fig, ax = plt.subplots(1, 1)
 
         # mark 1 year and 0.5 year
         year = 365.25
-        plt.axvline(x=year, ls='--', color='r', lw=3, alpha=0.6)
-        plt.axvline(x=year/2., ls='--', color='r', lw=3, alpha=0.6)
+        ax.axvline(x=year, ls='--', color='r', lw=3, alpha=0.6)
+        # ax.axvline(x=year/2., ls='--', color='r', lw=3, alpha=0.6)
         # plt.axvline(x=year/3., ls='--', color='r', lw=3, alpha=0.6)
 
         # mark the timespan of the data
-        plt.axvline(x=self.data[:,0].ptp(), ls='--', color='b', lw=4, alpha=0.5)
+        ax.axvline(x=self.data[:,0].ptp(), ls='--', color='b', lw=3, alpha=0.5)
 
         # by default, 100 bins in log between 0.1 and 1e7
         if bins is None:
             bins = 10 ** np.linspace(np.log10(1e-1), np.log10(1e7), 100)
 
-        plt.hist(T, bins=bins, alpha=0.5)
+        ax.hist(T, bins=bins, alpha=0.5)
 
-        plt.xscale("log")
-        plt.xlabel(r'(Period/days)')
-        plt.ylabel('Number of Posterior Samples')
-        plt.show()
+        ax.legend(['1 year', 'timespan'])
+        ax.set(xscale="log",
+               xlabel=r'(Period/days)',
+               ylabel='Number of Posterior Samples',
+               title='Posterior distribution for the orbital period(s)')
+        # plt.show()
 
 
     def make_plot3(self, points=True):
@@ -463,16 +472,17 @@ class KimaResults(object):
 
         if self.log_period:
             T = np.exp(self.T)
-            print('exponentiating period!')
+            # print('exponentiating period!')
         else:
             T = self.T
+
         A, E = self.A, self.E
 
 
         _, (ax1, ax2) = plt.subplots(2, 1, sharex=True)
 
         if points:
-            ax1.loglog(T, A, '.', markersize=1)
+            ax1.loglog(T, A, '.', markersize=2)
         else:
             ax1.hexbin(T, A, gridsize=50, 
                        bins='log', xscale='log', yscale='log',
[…]
