### 2018-06-12 8243b68
ditch argparse; new options for kima-showresults, related to #38; need to update docs
