### 2018-06-12 11574a9
add readme for BL2009 example; related to #36
