### 2018-06-12 353175e
create readme for the many_planets example
