### 2018-06-12 df3955b
add extra option to showresults; better check if in interactive interpreter
