### 2018-06-12 aaf152d
improve slightly makefile, make it easier to add new examples
