### 2018-06-12 c0f26c8
add new example; related to #37
