### 2018-06-12 918f5d7
fix some issues with skip and units; fixes #42; add histogram of extra sigma
