### 2018-06-12 20b6d41
improve makefile; add 51Peg example to makefile
