### 2018-06-12 j-faria/kima#44 @j-faria
Create a getting started guide going through the 51Peg example
Response to the JOSS referee (issue [here](https://github.com/openjournals/joss-reviews/issues/487))
