+++ README.md
@@ -11,6 +11,15 @@
 Clone with `--recursive` to get the submodules.  
 Running `make` will hopefully work. Need a fairly recent version of g++ (one that accepts `-std=c++11`)
 
+
+```bash
+git clone --recursive https://github.com/j-faria/kima.git
+cd kima
+make
+python setup.py install
+```
+
+
 See [the wiki](https://github.com/j-faria/kima/wiki) for more information and documentation.
 
 
