### 2018-06-14 b10bc86
move utility functions to utils.py
