### 2018-06-14 413db68
new API on all examples; simplify a bit
