+++ examples/CoRoT7/kima_setup.cpp
@@ -1,9 +1,7 @@
-#include <iostream>
 #include "DNest4.h"
 #include "Data.h"
 #include "RVmodel.h"
 
-using namespace std;
 using namespace DNest4;
 
 /* priors */
@@ -18,10 +16,7 @@ const bool trend = false;
 
 // options for the model
 // 
-RVmodel::RVmodel()
-    :planets(5, 5, false, RVConditionalPrior())
-    ,mu(Data::get_instance().N())
-    ,C(Data::get_instance().N(), Data::get_instance().N())
+RVmodel::RVmodel():fix(false),npmax(5)
 {
     double ymin = Data::get_instance().get_y_min();
     double ymax = Data::get_instance().get_y_max();
