### 2018-06-15 j-faria/kima#44 @j-faria
[Done](https://github.com/j-faria/kima/wiki/Getting-started)
