### 2018-06-15 1ab1920
update paper, citations
