+++ paper/paper.bib
@@ -99,3 +99,19 @@
   pages = {066001},
   file = {/home/joao/phd/papers/Fischer_etal_2016_2.pdf}
 }
+
+
+@article{Skilling2006,
+author = "Skilling, John",
+doi = "10.1214/06-BA127",
+fjournal = "Bayesian Analysis",
+journal = "Bayesian Analysis",
+month = "12",
+number = "4",
+pages = "833--859",
+publisher = "International Society for Bayesian Analysis",
+title = "Nested sampling for general Bayesian computation",
+url = "https://doi.org/10.1214/06-BA127",
+volume = "1",
+year = "2006"
+}
