### 2018-06-15 4c21398
change priors of 51 Peg example
