### 2018-06-15 #487 @j-faria
Hi @nespinoza,
I think I have now addressed all the issues raised in the review.

- There is a new example with data of 51 Peg.
  To go through this example step-by-step I created a new [getting started guide](https://github.com/j-faria/kima/wiki/Getting-started), which is included in the documentation.
- The `kima-showresults` script was changed to be a bit more intuitive
  and the [docs](https://github.com/j-faria/kima/wiki/Analysis-of-results) are now more clear, in my opinion.
- The new `kima-showresults rv` argument, in particular, addresses the question of plotting posterior realizations of the model over the radial-velocity measurements
- The documentation was also expanded and made clear in other parts.
  In particular, I included [a description](https://github.com/j-faria/kima/wiki/Input-data) of the type of input data *kima* expects,
  and expanded the pages about how to [contribute](https://github.com/j-faria/kima/wiki/Contribute) and how to [get support](https://github.com/j-faria/kima/wiki/Troubleshooting).
- There are new README files in each example
  which can, in the future, be expanded with more details about the results of the analysis.
- I changed slightly the paper and added citations to justify some of the claims.

Hopefully these changes make the package ready for publication. Please let me know
Allow me to once again acknowledge your contribution as referee, as I really think the changes brought up by your review improved *kima* substantially.

Regarding the **Version** and the github release, I think I'll need @arfon's help on this.
My idea was to release v1.0 once the submission is accepted.

Cheers,
João
