### 2018-06-18 #487 @nespinoza
Dear @j-faria,

 Thank you very much for considering the points made in the review. I have gone through all of them, including running the 51 Peg example, and everything runs smoothly!

 I have no more comments to give; I think the software is in its final form in terms of a JOSS publication and, @arfon, @j-faria, I'm happy to recommend `kima` for publication once (1) the final PDF is generated, (2) the first version is released (I suppose we need @arfon to double-check this). @j-faria, as before, congratulations on coding `kima`; it will be very useful for many of us working on exoplanets, and is a tool that definitely covers huge needs in the area of exoplanet discovery and characterizartion!

 Before finishing the review, one (optional, of course, as I have already recommended the software for publication) suggestion for @j-faria on `kima`: an explaination of each possible column that the `posterior_sample.txt` that `kima` generates for the  posterior samples of the model would be great. I know this is obvious for most of us working on exoplanets (e.g., `K` is obviously the RV semi-amplitude, `P` the period, etc.), but it might not be for everyone!

 All the best,
Néstor
