### 2018-06-18 #487 @j-faria
Thank you, @nespinoza!
The PDF is now generated (everything looks good I think). I will wait for instructions from @arfon on what to do next. And in the meantime I will add a description of the output files to the docs, I can probably do it before finishing the release.
