### 2018-06-18 #487 @arfon
@j-faria - please make the new release and then create an archive of this release with Zenodo or figshare. Once you have done this, please let me know what the DOI of this archive is here.
