### 2018-06-19 5d0f2b1
remove colors from make, RIP; fixes #43; cleanup
