### 2018-06-19 #487 @j-faria
Hi @arfon, here's the DOI from figshare: [10.6084/m9.figshare.6615350](https://doi.org/10.6084/m9.figshare.6615350)
