### 2018-06-21 #487 @j-faria
Thank you @barentsen and @nespinoza for the reviews, they both increased the quality of *kima* substantially. Thank you @arfon also for your editorial effort and help.
(and thanks @whedon :)
I'm very happy to see this published!
